New calibration parameters from 2/4/25
manual cherry-pick of commit 3f4981ef6a68e299e74273a9673d10040b1d70f2

This includes the new calibration parameters as of 2/4/25, but without
the extra __pycache__ files and without some conflicting changes to user
interface.

All of these paramters would ideally go in a single place, with just one
copy for all code to use. Either a single python file with all the
parameters, or it could be a json file or similar?