Warn if joint out of range, allow negatives
The base limits are set to range from 0 to 1000 clicks.
If the base is rotated a bit further by hand, or strays a bit further by
momentum, then the click value can underflow to 65535 or similar values.

This patch makes three changes to address this:
- warn on the console if the click value is outside the expected range
- if so, force the click value into the expected range
- use signed integers for clicks (work in progress, not yet complete) so
  that 65535 is converted to -1, etc.

Also implement some improvements to the front end scripts.