Fix documentation bug.
The description of packet length was wrong, it didn't match the code in
our files, or code in various examples online. The official
documentation is very unclear on the exact length calculations.

Also remove an extra write_out that is not needed, it was being sent to
the simulation during read operations even when the arm was also used,
but the simulation isn't used for reads in that case.