multi-joint commands
Ability to query position of multiple joints simultaneously, in one USB
packet. Similarly, ability to set position of multiple joints
simultaneously, in one USB packet.

This is meant to make the USB part of the protocol a little more
efficient, since the command packets to the arm are able to carry an
array of positions for different servos, not just a single position for
one servo.