game and direct-movement prototype, working
This implements a wildly different approach to movement, not using the
inverse jacobian at all, and makes no attempt to move the end-effector
along a linear pat.  Instead, it directly solves for the joint angles
needed to reach or approach the target position, and moves the arm in a
single servo command to that position.

This also makes a few bug fixes to the simulator, so the gripper and
hand match the physical arm, and fixes the simulator animation speed to
be more in line with the physical arm.